Refactoring and bugfixes.
Changed layout of directories in `ux` mode for DLC packages.
Updated version strings.
